django-simple-api/requirements.txt
django==5.2.5
gunicorn==23.0.0
kafka-python==2.0.2